// dmgcc/10_Landing_Page/src/saga/Types/gccvertical.types.ts
export enum GccVerticalTypes {
    //cost centre master
    GET_CAPABILITIES_REQUEST = 'GET_CAPABILITIES_REQUEST',
    GET_CAPABILITIES_SUCCESS = 'GET_CAPABILITIES_SUCCESS',
    GET_CAPABILITIES_ERROR = 'GET_CAPABILITIES_ERROR',
    GET_MANUFACTURING_ENGINEER_REQUEST = 'GET_MANUFACTURING_ENGINEER_REQUEST',
    GET_MANUFACTURING_ENGINEER_SUCCESS = 'GET_MANUFACTURING_ENGINEER_SUCCESS',
    GET_MANUFACTURING_ENGINEER_ERROR = 'GET_MANUFACTURING_ENGINEER_ERROR',
    GET_MANUFACTURING_ENGINEER_SUB_CONTENT_REQUEST = 'GET_MANUFACTURING_ENGINEER_SUB_CONTENT_REQUEST',
    GET_MANUFACTURING_ENGINEER_SUB_CONTENT_SUCCESS = 'GET_MANUFACTURING_ENGINEER_SUB_CONTENT_SUCCESS',
    GET_MANUFACTURING_ENGINEER_SUB_CONTENT_ERROR = 'GET_MANUFACTURING_ENGINEER_SUB_CONTENT_ERROR',
    GET_PRODUCT_ENGINEER_REQUEST = 'GET_PRODUCT_ENGINEER_REQUEST',
    GET_PRODUCT_ENGINEER_SUCCESS = 'GET_PRODUCT_ENGINEER_SUCCESS',
    GET_PRODUCT_ENGINEER_ERROR = 'GET_PRODUCT_ENGINEER_ERROR',
    GET_PRODUCT_ENGINEER_SUB_CONTENT_REQUEST = 'GET_PRODUCT_ENGINEER_SUB_CONTENT_REQUEST',
    GET_PRODUCT_ENGINEER_SUB_CONTENT_SUCCESS = 'GET_PRODUCT_ENGINEER_SUB_CONTENT_SUCCESS',
    GET_PRODUCT_ENGINEER_SUB_CONTENT_ERROR = 'GET_PRODUCT_ENGINEER_SUB_CONTENT_ERROR',
    GET_CUSTOMER_SERVICES_REQUEST = 'GET_CUSTOMER_SERVICES_REQUEST',
    GET_CUSTOMER_SERVICES_SUCCESS = 'GET_CUSTOMER_SERVICES_SUCCESS',
    GET_CUSTOMER_SERVICES_ERROR = 'GET_CUSTOMER_SERVICES_ERROR',
    GET_CUSTOMER_SERVICES_SUB_CONTENT_REQUEST = 'GET_CUSTOMER_SERVICES_SUB_CONTENT_REQUEST',
    GET_CUSTOMER_SERVICES_SUB_CONTENT_SUCCESS = 'GET_CUSTOMER_SERVICES_SUB_CONTENT_SUCCESS',
    GET_CUSTOMER_SERVICES_SUB_CONTENT_ERROR = 'GET_CUSTOMER_SERVICES_SUB_CONTENT_ERROR',
    GET_INFORMATION_RECHNOLOGY_REQUEST = 'GET_INFORMATION_RECHNOLOGY_REQUEST',
    GET_INFORMATION_RECHNOLOGY_SUCCESS = 'GET_INFORMATION_RECHNOLOGY_SUCCESS',
    GET_INFORMATION_RECHNOLOGY_ERROR = 'GET_INFORMATION_RECHNOLOGY_ERROR',
    GET_FINANCE_CONTROLLING_REQUEST = 'GET_FINANCE_CONTROLLING_REQUEST',
    GET_FINANCE_CONTROLLING_SUCCESS = 'GET_FINANCE_CONTROLLING_SUCCESS',
    GET_FINANCE_CONTROLLING_ERROR = 'GET_FINANCE_CONTROLLING_ERROR',
    GET_FINANCE_CONTROLLING_SUB_CONTENT_REQUEST = 'GET_FINANCE_CONTROLLING_SUB_CONTENT_REQUEST',
    GET_FINANCE_CONTROLLING_SUB_CONTENT_SUCCESS = 'GET_FINANCE_CONTROLLING_SUB_CONTENT_SUCCESS',
    GET_FINANCE_CONTROLLING_SUB_CONTENT_ERROR = 'GET_FINANCE_CONTROLLING_SUB_CONTENT_ERROR',
    GET_COST_ENGINEERING_REQUEST = 'GET_COST_ENGINEERING_REQUEST',
    GET_COST_ENGINEERING_SUCCESS = 'GET_COST_ENGINEERING_SUCCESS',
    GET_COST_ENGINEERING_ERROR = 'GET_COST_ENGINEERING_ERROR',
    GET_COST_ENGINEERING_SUB_CONTENT_REQUEST = 'GET_COST_ENGINEERING_SUB_CONTENT_REQUEST',
    GET_COST_ENGINEERING_SUB_CONTENT_SUCCESS = 'GET_COST_ENGINEERING_SUB_CONTENT_SUCCESS',
    GET_COST_ENGINEERING_SUB_CONTENT_ERROR = 'GET_COST_ENGINEERING_SUB_CONTENT_ERROR',
    GET_HUMAN_RESOURCES_REQUEST = 'GET_HUMAN_RESOURCES_REQUEST',
    GET_HUMAN_RESOURCES_SUCCESS = 'GET_HUMAN_RESOURCES_SUCCESS',
    GET_HUMAN_RESOURCES_ERROR = 'GET_HUMAN_RESOURCES_ERROR',
    GET_HUMAN_RESOURCES_SUB_CONTENT_REQUEST = 'GET_HUMAN_RESOURCES_SUB_CONTENT_REQUEST',
    GET_HUMAN_RESOURCES_SUB_CONTENT_SUCCESS = 'GET_HUMAN_RESOURCES_SUB_CONTENT_SUCCESS',
    GET_HUMAN_RESOURCES_SUB_CONTENT_ERROR = 'GET_HUMAN_RESOURCES_SUB_CONTENT_ERROR',
    GET_QUALITY_MANAGEMENT_REQUEST = 'GET_QUALITY_MANAGEMENT_REQUEST',
    GET_QUALITY_MANAGEMENT_SUCCESS = 'GET_QUALITY_MANAGEMENT_SUCCESS',
    GET_QUALITY_MANAGEMENT_ERROR = 'GET_QUALITY_MANAGEMENT_ERROR',
    GET_QUALITY_MANAGEMENT_SUB_CONTENT_REQUEST = 'GET_QUALITY_MANAGEMENT_SUB_CONTENT_REQUEST',
    GET_QUALITY_MANAGEMENT_SUB_CONTENT_SUCCESS = 'GET_QUALITY_MANAGEMENT_SUB_CONTENT_SUCCESS',
    GET_QUALITY_MANAGEMENT_SUB_CONTENT_ERROR = 'GET_QUALITY_MANAGEMENT_SUB_CONTENT_ERROR',
    GET_LEAN_PROCESS_REQUEST = 'GET_LEAN_PROCESS_REQUEST',
    GET_LEAN_PROCESS_SUCCESS = 'GET_LEAN_PROCESS_SUCCESS',
    GET_LEAN_PROCESS_ERROR = 'GET_LEAN_PROCESS_ERROR',
    GET_LEAN_PROCESS_SUB_CONTENT_REQUEST = 'GET_LEAN_PROCESS_SUB_CONTENT_REQUEST',
    GET_LEAN_PROCESS_SUB_CONTENT_SUCCESS = 'GET_LEAN_PROCESS_SUB_CONTENT_SUCCESS',
    GET_LEAN_PROCESS_SUB_CONTENT_ERROR = 'GET_LEAN_PROCESS_SUB_CONTENT_ERROR',
    GET_SUPPLIER_MANAGEMENT_REQUEST = 'GET_SUPPLIER_MANAGEMENT_REQUEST',
    GET_SUPPLIER_MANAGEMENT_SUCCESS = 'GET_SUPPLIER_MANAGEMENT_SUCCESS',
    GET_SUPPLIER_MANAGEMENT_ERROR = 'GET_SUPPLIER_MANAGEMENT_ERROR',
    GET_SUPPLIER_MANAGEMENT_SUB_CONTENT_REQUEST = 'GET_SUPPLIER_MANAGEMENT_SUB_CONTENT_REQUEST',
    GET_SUPPLIER_MANAGEMENT_SUB_CONTENT_SUCCESS = 'GET_SUPPLIER_MANAGEMENT_SUB_CONTENT_SUCCESS',
    GET_SUPPLIER_MANAGEMENT_SUB_CONTENT_ERROR = 'GET_SUPPLIER_MANAGEMENT_SUB_CONTENT_ERROR',
}